+++ src/main/java/com/zhongmei/yunfu/controller/PushPlanNewDishController.java
@@ -116,6 +116,9 @@ public class PushPlanNewDishController extends BaseController {
 
     @RequestMapping("/delete")
     public String deleteNewDishPushPlan(Model model, Long id) {
+        if(id == null){
+            return "fail";
+        }
         boolean result = mPushNewDishService.deleteNewDishPushPlan(id);
         return "success";
     }
